refactor: CodeRabbit 2차 리뷰 반영 #89
- config.ts를 types.ts로 합쳤다. 컴포넌트는 UI 계약 타입만 봐야 하는데
  `BillingConfig`만 따로 살고 있었다 — types.ts가 스스로 "UI 계약"이라 적고 있어
  둘로 나눌 이유가 없었다(CLAUDE.md §Mock → Live 격리막)
- checkout-panel 215줄 → 156줄. Row·Agreement를 checkout-rows로 분리.
  common으로 올리지 않았다 — 쓰는 데가 이 화면 하나뿐이다
- 요일 오류 둘. 2026-08-02는 일요일인데 `8월 2일(토)`, CLAUDE.md 날짜 예시도
  `8월 5일(화)`인데 실제로는 수요일이다
