+++ CLAUDE.md
@@ -119,7 +119,7 @@ app/
   - ⚠️ **위 폭은 목표치이지 고정값이 아니다.** `w-[1440px]` 대신 `mx-auto max-w-[1440px] px-8`, absolute 대신 flex/grid, 표는 `overflow-x-auto`로 감싼다. 사이드바는 컴포넌트로 분리(모바일은 Sheet).
   - 반응형 전면 구현은 지금 안 한다. 대상 화면 선별은 디자인 확정 후.
 - 폼 2열(`FormRow`) · 제출 버튼 하단우측 · 로딩=스켈레톤 · 모션 100/150/250ms · 숫자 `tabular-nums`
-- **카피:** **~합니다체**(2026-08-04 변경) · 날짜 `8월 5일(화)` · 역할 워딩은 영어
+- **카피:** **~합니다체**(2026-08-04 변경) · 날짜 `8월 5일(수)` · 역할 워딩은 영어
   - 사내 도구이고 **돈·권한·기록이 걸린 화면**이라 친근한 말투가 오히려 가볍게 읽힌다. `들어올 수 없어요` → `접근할 수 없습니다`.
   - 명령은 **`~해 주세요`** 를 쓴다. `~하십시오`는 딱딱해서 안 쓴다.
   - ⚠️ 옛 화면에 `~해요체`가 남아 있으면 그건 이 변경 전 것이다.
