+++ src/features/billing/components/checkout-panel.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { ArrowRight, Check, RefreshCw } from "lucide-react";
-import type { ReactNode } from "react";
+import { ArrowRight, RefreshCw } from "lucide-react";
 
 import { ZLogo } from "@/components/icons/z-logo";
-import { cn } from "@/lib/utils";
 
-import type { BillingConfig } from "../config";
 import { calculatePrice, formatWon } from "../pricing";
+import type { BillingConfig } from "../types";
+import { Agreement, Row } from "./checkout-rows";
 
 interface CheckoutPanelProps {
   config: BillingConfig;
@@ -151,64 +150,5 @@ export function CheckoutPanel({
           */}
       </aside>
     </div>
-  );
-}
-
-function Row({ label, value, isMuted }: { label: string; value: string; isMuted?: boolean }) {
-  return (
-    <div className="flex items-center justify-between">
-      <dt
-        className={
-          isMuted
-            ? "text-muted-foreground/70 text-[11px] leading-4"
-            : "text-muted-foreground text-[13px] leading-5"
-        }
-      >
-        {label}
-      </dt>
-      <dd
-        className={
-          isMuted
-            ? "text-muted-foreground/70 text-[11px] leading-4 tabular-nums"
-            : "text-[13px] leading-5 tabular-nums"
-        }
-      >
-        {value}
-      </dd>
-    </div>
-  );
-}
-
-function Agreement({
-  checked,
-  onChange,
-  children,
-}: {
-  checked: boolean;
-  onChange: (next: boolean) => void;
-  children: ReactNode;
-}) {
-  return (
-    <label className="flex cursor-pointer items-start gap-2.5">
-      {/*
-        ⚠️ 브라우저 기본 체크박스는 OS마다 다르게 생겨 이 카드만 마감이 덜 된 것처럼 보인다.
-           상자는 우리가 그리고, 진짜 `input`은 숨겨서 키보드·스크린리더는 그대로 쓴다.
-      */}
-      <input
-        type="checkbox"
-        checked={checked}
-        onChange={(event) => onChange(event.target.checked)}
-        className="peer sr-only"
-      />
-      <span
-        className="border-input bg-background peer-checked:bg-foreground peer-checked:border-foreground peer-focus-visible:ring-ring text-background mt-[1px] flex size-[18px] shrink-0 items-center justify-center rounded-md border transition-colors peer-focus-visible:ring-2 peer-focus-visible:ring-offset-2"
-        aria-hidden
-      >
-        <Check className={cn("size-3 transition-opacity", checked ? "opacity-100" : "opacity-0")} />
-      </span>
-      <span className="text-muted-foreground text-[12px] leading-[18px] break-keep">
-        {children}
-      </span>
-    </label>
   );
 }
